Finalizado parte do Aluno

+++ Promart/Models/AlunoVinculo.cs
@@ -13,10 +13,10 @@ namespace Promart.Models
     {
         public int IdAluno { get; set; }
         public string? NomeFamiliar { get; set; }
-        public ushort Idade { get; set; }
+        public int Idade { get; set; }
         public string? Parentesco { get; set; }
         public string? Ocupacao { get; set; }
         public string? Escolaridade { get; set; }          
-        public ushort Renda { get; set; }
+        public int Renda { get; set; }
     }
 }
